♻️refactor: event 주인인지에 대한 검증로직처리 및 변수명 정정
