+++ src/main/java/com/project/backend/domain/common/reminder/dto/PlanChanged.java
@@ -1,4 +1,4 @@
-package com.project.backend.domain.event.dto;
+package com.project.backend.domain.common.reminder.dto;
 
 import com.project.backend.domain.reminder.enums.ChangeType;
 import com.project.backend.domain.reminder.enums.TargetType;
@@ -12,7 +12,7 @@ public record PlanChanged(
         TargetType targetType,
         Long memberId,
         String title,
-        Boolean isrRecurring,
+        Boolean isRecurring,
         // 생성 / 수정에서만 의미 있음
         LocalDateTime occurrenceTime,
         ChangeType changeType
